Validate image buffer boundaries

+++ src/platform/win32/ImgSettings.h
@@ -2,9 +2,62 @@
 
 #include <Windows.h>
 #include <shlobj.h>
-#include <string>
 #include <iomanip>
 #include <sstream>
+#include <stdexcept>
+#include <string>
+
+enum class ImgSettingsStatus
+{
+    TempPathFailed,
+    GuidCreationFailed,
+    TempDirectoryCreationFailed
+};
+
+class ImgSettingsError final : public std::runtime_error
+{
+  public:
+    ImgSettingsError(ImgSettingsStatus status, HRESULT system_error)
+        : std::runtime_error(BuildMessage(status, system_error)), status_(status), system_error_(system_error)
+    {
+    }
+
+    ImgSettingsStatus status() const
+    {
+        return status_;
+    }
+
+    HRESULT system_error() const
+    {
+        return system_error_;
+    }
+
+  private:
+    static std::string BuildMessage(ImgSettingsStatus status, HRESULT system_error)
+    {
+        const char* operation = "Unknown";
+        switch (status)
+        {
+        case ImgSettingsStatus::TempPathFailed:
+            operation = "GetTempPath";
+            break;
+        case ImgSettingsStatus::GuidCreationFailed:
+            operation = "CoCreateGuid";
+            break;
+        case ImgSettingsStatus::TempDirectoryCreationFailed:
+            operation = "CreateDirectory";
+            break;
+        }
+
+        std::stringstream message;
+        message << "ImgSettings." << operation << "() failed with error 0x" << std::hex << std::setw(8)
+                << std::setfill('0') << std::uppercase << static_cast<unsigned long>(system_error);
+        return message.str();
+    }
+
+    ImgSettingsStatus status_;
+    HRESULT system_error_;
+};
 
 class ImgSettings
 {
@@ -43,12 +96,18 @@ class ImgSettings
 
 inline void ImgSettings::InitializeTempPath()
 {
-    TCHAR temppathbuffer[MAX_PATH + 1];
+    TCHAR temppathbuffer[MAX_PATH + 1]{};
 
     const auto pathlen = GetTempPath(MAX_PATH, temppathbuffer);
-    if (pathlen > MAX_PATH || pathlen == 0)
+    if (pathlen == 0)
     {
-        // TODO: handle error.
+        const auto error = GetLastError();
+        throw ImgSettingsError(ImgSettingsStatus::TempPathFailed,
+                               HRESULT_FROM_WIN32(error == ERROR_SUCCESS ? ERROR_PATH_NOT_FOUND : error));
+    }
+    if (pathlen >= MAX_PATH)
+    {
+        throw ImgSettingsError(ImgSettingsStatus::TempPathFailed, HRESULT_FROM_WIN32(ERROR_INSUFFICIENT_BUFFER));
     }
 
     std::wstring temppath = temppathbuffer;
@@ -58,39 +117,36 @@ inline void ImgSettings::InitializeTempPath()
         temppath += L'\\';
     }
 
-    std::wstring imgtemppath = temppath;
-    std::wstring testpath;
-    WIN32_FIND_DATA finddata{};
-    HANDLE find{};
-    BOOL folderexists = TRUE;
-
-    do
+    for (;;)
     {
-        GUID guid;
-        if (CoCreateGuid(&guid) != S_OK)
+        GUID guid{};
+        const auto guidresult = CoCreateGuid(&guid);
+        if (FAILED(guidresult))
         {
-            // TODO: handle error.
+            throw ImgSettingsError(ImgSettingsStatus::GuidCreationFailed, guidresult);
         }
 
         std::wstringstream wss;
         wss << std::setw(8) << std::setfill(L'0') << std::uppercase << std::hex << guid.Data1;
-        testpath = temppath + wss.str();
-
-        find = FindFirstFile(testpath.c_str(), &finddata);
-        if (find == INVALID_HANDLE_VALUE)
+        const auto testpath = temppath + wss.str();
+        if (CreateDirectory(testpath.c_str(), nullptr))
         {
-            CreateDirectory(testpath.c_str(), NULL);
-            folderexists = FALSE;
             temppath_ = testpath;
+            return;
         }
-        else
+
+        const auto error = GetLastError();
+        if (error != ERROR_ALREADY_EXISTS)
         {
-            FindClose(find);
+            throw ImgSettingsError(ImgSettingsStatus::TempDirectoryCreationFailed, HRESULT_FROM_WIN32(error));
         }
-    } while (folderexists);
+    }
 }
 
 inline void ImgSettings::DeleteTempPath()
 {
-    RemoveDirectory(temppath_.c_str());
+    if (!temppath_.empty())
+    {
+        RemoveDirectory(temppath_.c_str());
+    }
 }
